+++ Pocket Casts TV App/UI/Podcasts/EpisodeRow.swift
@@ -312,5 +312,5 @@ extension View {
 #Preview {
     EpisodeRowWithActions(model: EpisodeRowViewModel(episode: MockData.makeStubEpisodes().first!, podcast: MockData.makeStubPodcasts().first!))
     .environment(AppCoordinator())
-    .environment(MainTabRouter())
+    .environment(MainTabViewModel())
 }